Added HTML parser integration and tests
